fixed issue #3
