+++ src/json/cover.rs
@@ -1,9 +1,10 @@
+use http_cache_reqwest::{CACacheManager, Cache, CacheMode, HttpCache};
+use reqwest_middleware::ClientBuilder;
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
-use crate::client::util::construct_url;
 use crate::prelude::*;
-use crate::util::*;
+use crate::util::client::construct_url;
 use crate::{unwrap_api_results, uuid_or_err};
 
 ////////////////////////////////////////////////////////////////
@@ -66,8 +67,22 @@ impl Client<Cover> {
     }
 
     pub async fn get(self) -> Result<Cover, ClientError> {
+        let client = ClientBuilder::new(reqwest::Client::new())
+            .with(Cache(HttpCache {
+                mode: CacheMode::Default,
+                manager: CACacheManager {
+                    path: if let Some(mut path) = dirs::cache_dir() {
+                        path.push("mangadex_api-cacache");
+                        path
+                    } else {
+                        std::path::PathBuf::from("./mangadex_api-cacache")
+                    },
+                },
+                options: None,
+            }))
+            .build();
         let uuid = uuid_or_err!(self.get_uuid()).unwrap();
-        let res = CLIENT
+        let res = client
             .get(construct_url(format!("/cover/{uuid}"), None))
             .send()
             .await
@@ -111,7 +126,21 @@ impl Client<CoverList> {
     }
 
     pub async fn get(self) -> Result<CoverList, ClientError> {
-        let res = CLIENT
+        let client = ClientBuilder::new(reqwest::Client::new())
+            .with(Cache(HttpCache {
+                mode: CacheMode::Default,
+                manager: CACacheManager {
+                    path: if let Some(mut path) = dirs::cache_dir() {
+                        path.push("mangadex_api-cacache");
+                        path
+                    } else {
+                        std::path::PathBuf::from("./mangadex_api-cacache")
+                    },
+                },
+                options: None,
+            }))
+            .build();
+        let res = client
             .get(construct_url(format!("/cover"), None))
             .send()
             .await
