+++ tests/client.rs
@@ -6,7 +6,7 @@ async fn manga() {
     let res = Client::<Manga>::new(uuid!("c288b108-5162-4065-aa3a-5857ec38c8d9"))
         .get()
         .await;
-    assert!(res.is_ok(), "{:#?}", res.unwrap())
+    assert!(res.is_ok(), "{:#?}", res.unwrap_err())
 }
 
 #[tokio::test]
@@ -14,13 +14,13 @@ async fn manga_feed() {
     let res = Client::<MangaFeed>::new(uuid!("c288b108-5162-4065-aa3a-5857ec38c8d9"))
         .get()
         .await;
-    assert!(res.is_ok(), "{:#?}", res.unwrap())
+    assert!(res.is_ok(), "{:#?}", res.unwrap_err())
 }
 
 #[tokio::test]
 async fn manga_list() {
     let res = Client::<MangaList>::new().get().await;
-    assert!(res.is_ok(), "{:#?}", res.unwrap())
+    assert!(res.is_ok(), "{:#?}", res.unwrap_err())
 }
 
 #[tokio::test]
@@ -28,7 +28,7 @@ async fn athome_server() {
     let res = Client::<AtHomeServer>::new(uuid!("9e989fd3-72bf-4f99-a24c-114166be880a"))
         .get()
         .await;
-    assert!(res.is_ok(), "{:#?}", res.unwrap())
+    assert!(res.is_ok(), "{:#?}", res.unwrap_err())
 }
 
 #[tokio::test]
@@ -36,13 +36,13 @@ async fn author() {
     let res = Client::<Author>::new(uuid!("c6c278e1-268b-4b7b-84ec-3289bd0c08f0"))
         .get()
         .await;
-    assert!(res.is_ok(), "{:#?}", res.unwrap())
+    assert!(res.is_ok(), "{:#?}", res.unwrap_err())
 }
 
 #[tokio::test]
 async fn author_list() {
     let res = Client::<AuthorList>::new().get().await;
-    assert!(res.is_ok(), "{:#?}", res.unwrap())
+    assert!(res.is_ok(), "{:#?}", res.unwrap_err())
 }
 
 #[tokio::test]
@@ -60,13 +60,13 @@ async fn cover() {
         .uuid;
 
     let res = Client::<Cover>::new(uuid).get().await;
-    assert!(res.is_ok(), "{:#?}", res.unwrap())
+    assert!(res.is_ok(), "{:#?}", res.unwrap_err())
 }
 
 #[tokio::test]
 async fn cover_list() {
     let res = Client::<CoverList>::new().get().await;
-    assert!(res.is_ok(), "{:#?}", res.unwrap())
+    assert!(res.is_ok(), "{:#?}", res.unwrap_err())
 }
 
 #[tokio::test]
@@ -74,11 +74,11 @@ async fn chapter() {
     let res = Client::<Chapter>::new(uuid!("9e989fd3-72bf-4f99-a24c-114166be880a"))
         .get()
         .await;
-    assert!(res.is_ok(), "{:#?}", res.unwrap())
+    assert!(res.is_ok(), "{:#?}", res.unwrap_err())
 }
 
 #[tokio::test]
 async fn chapter_list() {
     let res = Client::<ChapterList>::new().get().await;
-    assert!(res.is_ok(), "{:#?}", res.unwrap())
+    assert!(res.is_ok(), "{:#?}", res.unwrap_err())
 }
